Improve worker termination; Avoid duplicated opencv/ort fetches
